+++ API/CMakeLists.txt
@@ -1,7 +1,9 @@
-cmake_minimum_required(VERSION 3.22)
+cmake_minimum_required(VERSION 3.16)
 project(plugin_api)
 
-add_library(plugin_api INTERFACE include/plugin_api/motion_planner_interface.hpp)
+set(HEADERS include/plugin_api/motion_planner_interface.h)
+
+add_library(plugin_api INTERFACE include/plugin_api/motion_planner_interface.h)
 
 target_compile_features(plugin_api INTERFACE cxx_std_17)
 target_include_directories(
@@ -14,8 +16,7 @@ install(
   LIBRARY DESTINATION lib
   ARCHIVE DESTINATION lib
   RUNTIME DESTINATION bin
-  INCLUDES
-  DESTINATION include)
+  INCLUDES DESTINATION include)
 install(DIRECTORY include/ DESTINATION include)
 
 install(
@@ -23,11 +24,3 @@ install(
   FILE plugin_apiTargets.cmake
   NAMESPACE plugin_api::
   DESTINATION lib/cmake/plugin_api)
-
-# Install
-include(CMakePackageConfigHelpers)
-
-# Guard for those whole include this package directly
-if(NOT CMAKE_PROJECT_NAME STREQUAL PROJECT_NAME)
-  return()
-endif()
