+++ cmake/CppConPluginExampleConfig.cmake
@@ -1,0 +1,2 @@
+include(CMakeFindDependencyMacro)
+include(${CMAKE_CURRENT_LIST_DIR}/CppConPluginExampleTargets.cmake)
