+++ core/plugin_manager/meta_object.hpp
