+++ core/plugin_manager/rosful_class_loader.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <class_loader.hpp>
+#include <class_loader.h>
 
 #pragma once
