+++ core/plugin_manager/shared_library.hpp
