+++ libraries/simple_motion_planner/cmake/simple_motion_plannerConfig.cmake
@@ -1,0 +1,2 @@
+include(CMakeFindDependencyMacro)
+include(${CMAKE_CURRENT_LIST_DIR}/simple_motion_plannerTargets.cmake)
