+++ libraries/simple_motion_planner/simple_motion_planner.h
@@ -1,25 +1,25 @@
 #pragma once
  
 #include <iostream>
-#include "motion_planner_interface.hpp"
+#include "plugin_api/motion_planner_interface.h"
 
-class SimpleMotionPlanner : public motion_planner::IMotionPlanner
+class SimpleMotionPlanner : public IMotionPlanner
 {
 public:
 	SimpleMotionPlanner() = default;
 	~SimpleMotionPlanner() = default;
-
+ 
 	bool initialize() override
 	{
 		// Here is where MoveIt reads parameters and creates publishers/subscribers/services/actions
 		return true;
 	}
- 
-	std::vector<motion_planner::RobotState> plan(motion_planner::RobotState start, motion_planner::RobotState goal) override
+
+	std::vector<RobotState> plan(RobotState start, RobotState goal) override
 	{
 		// Insert a simple motion planner
 		std::cout << "I am a planning motion simply\n";
-		return std::vector<motion_planner::RobotState>{};
+		return std::vector<RobotState>{};
 	}
 };
